Arrumando erros de merge e adicionando DTOs de historico

+++ src/main/java/com/marmitech/Marmitech/Entity/Pedido.java
@@ -38,7 +38,7 @@ public class Pedido {
 
     private String dataPedido;
 
-    private Double valorTotal;
+    private BigDecimal valorTotal;
 
     @NotNull
     @NotBlank(message = "status não pode ser null ou vazio")
